Of course I can add multiple filetypes to a single pattern.
When working with phtml files, I want highlighting for HTML (and
HTML5), php, JavaScript, and CSS. I was setting the filetype to
php and getting close to the results I wanted, but the HTML5
elements were not correct.

The other issue was with JavaScript highlighting.

It's all fixed now. I had no clue you could chain filetypes
together.